Checkpointing -- hopefully no change to functionality. This includes the equations & code for several constraints in the new constraint format, but those equations are not yet being used for anything.

+++ auxiliary/tests/CoarseRNA/CoarseRNA.cpp
@@ -260,7 +260,7 @@ try // If anything goes wrong, an exception will be thrown.
     cout << "sub.ancestor=" << sub.getAncestorBody();
 //    cout << "  sub.terminalBodies=" << sub.getTerminalBodies() << endl;
 //    cout << "sub.allBodies=" << sub.getAllBodies() << endl;
-    for (SubtreeBodyId i(0); i < sub.getAllBodies().size(); ++i) {
+    for (SubtreeBodyId i(0); i < (int)sub.getAllBodies().size(); ++i) {
        cout << "sub.parent[" << i << "]=" << sub.getParentSubtreeBodyId(i);
 //       cout << "  sub.children[" << i << "]=" << sub.getChildSubtreeBodyIds(i) << endl;
     }
